requirements.txt
python-dotenv
black
pylint
pytest
pandas
matplotlib